Stop routing pointers to func_8004036C through s32 (#2500)
func_8004036C takes void *. Eight call sites hand it an integer instead, which
GCC reports as "makes pointer from integer without a cast". They are one class
in two shapes, and neither is load-bearing.

A redundant (s32) on something already a pointer. D_8009B188, D_8009B18C and
D_8009B3D8 are all declared u8 *, and p in func_800313E8.c is a u8 * local, so
the cast turns a pointer into an integer purely so the callee can turn it back.

A stored pointer read as *(s32 *). func_800391E4.c reads the slot at 0x2C of a
DuelEffect channel that way, while text_box_lifecycle.c and
text_box_build_step.c read the same record at the same offset as *(void **).
The tree already has a spelling for this operation; that one file just does not
use it. Both the guard and the call are converted, so the slot is not read two
ways on adjacent lines.

This is a batch on purpose, in the shape #2936 asked for: one callee, one
warning class, and the same review question at every site -- is the value a
pointer? By declaration, yes at all of them.

dialog_update_choice.c:79 is the eighth site and is deliberately left out. That
file has an open PR of mine, #3129, and editing it from two changes at once
tangles the queue for no benefit. Its warning is the one that survives here.

Warnings naming func_8004036C: eight to one, the remainder being that
deferred site.

make match reproduces the retail SHA; basic-types, check-metadata and
match-overlays pass.

+++ src/game/func_80020988.c
@@ -71,8 +71,8 @@ s32 func_80020988(void) {
 
     if ((gInput_wPad1Pressed &
          (PAD_BUTTON_CONFIRM_MASK | PAD_BUTTON_CANCEL)) != 0) {
-        func_8004036C((s32)D_8009B188);
-        func_8004036C((s32)D_8009B18C);
+        func_8004036C(D_8009B188);
+        func_8004036C(D_8009B18C);
         if ((gInput_wPad1Pressed & PAD_BUTTON_CANCEL) != 0) {
             return -1;
         }
